Refactor item creation and deletion logic in SolutionOrdersAPI by removing IsActive assignment from CreateItemHandler and adding null check in DeleteItemHandler. Update mapping configurations for CreateItemCommand and UpdateItemCommand to ignore unnecessary properties, ensuring cleaner data handling.

+++ SolutionOrdersAPI/SolutionOrdersAPI/Features/Items/Handlers/Commands/DeleteItemHandler.cs
@@ -12,6 +12,10 @@ public class DeleteItemHandler(IItemProvider itemProvider, IItemService itemServ
     public async Task<Unit> Handle(DeleteItemCommand request, CancellationToken cancellationToken)
     {
         var item = await itemProvider.GetItemByIdAsync(request.IdItem, false, cancellationToken);
+        
+        if (item == null)
+            throw new KeyNotFoundException($"Produkt o ID {request.IdItem} nie istnieje.");
+        
         logger.LogInformation("Usuwanie produktu ID: {IdItem}", request.IdItem);
         item.IsActive = false;
         await itemService.DeleteItem(item, cancellationToken);
